bookstore_requirements.txt
google-generativeai
python-dotenv
faiss-cpu
numpy
